pyproject.toml
[tool.poetry]
name = "tapas"
version = "0.1.0"
description = ""
authors = ["nanoastax <jakub.kramp@profil-software.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.111.0"
sqlmodel = "^0.0.19"
uvicorn = "^0.20.0"
alembic = "^1.9.3"
psycopg2-binary = "^2.9.5"
django-environ = "^0.9.0"
aiohttp = "^3.8.4"
requests = "^2.28.2"
python-multipart = "^0.0.7"
bcrypt = "^4.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.23.3"
pytest = "^7.2.2"
SQLAlchemy-Utils = "^0.41.1"
pytest-mock = "^3.10.0"
pytest-cov = "^4.0.0"
ruff = "^0.0.262"
selenium = "^4.9.0"
webdriver-manager = "^3.8.6"
beautifulsoup4 = "^4.12.2"
lxml = "^4.9.2"
pydantic-settings = "^2.3.4"

[tool.poetry.dev-dependencies]
pytest = "^7.2.2"
black = "^23.1.0"
pre-commit = "^3.1.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 110
exclude = '''
/(
    \.venv      # Exclude virtual environment directories
    | migrations    # Exclude build directories
)/
'''

[tool.ruff]
line-length = 110
extend-exclude = ['migrations/']